Use Doubles

+++ gart/src/main/kotlin/dev/oblac/gart/fluid/lbh/fluidBoltzmann.kt
@@ -3,30 +3,30 @@ package dev.oblac.gart.fluid.lbh
 
 @Suppress("t")
 class BoltzmannFluid(
-    private val overallVelocity: Float,     // wind speed
-    private val viscosity: Float,
+    private val overallVelocity: Double,     // wind speed
+    private val viscosity: Double,
     private val rows: Int,
     private val cols: Int
 ) {
     // discretized thermal velocities for each of the 9 possible directions
-    private val velocityHere: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocityUp: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocityDown: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocityRight: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocityLeft: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocityNorthWest: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocityNorthEast: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocitySouthWest: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val velocitySouthEast: Array<FloatArray> = Array(rows) { FloatArray(cols) }
+    private val velocityHere: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocityUp: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocityDown: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocityRight: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocityLeft: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocityNorthWest: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocityNorthEast: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocitySouthWest: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val velocitySouthEast: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
 
     private val solid: Array<BooleanArray> = Array(rows) { BooleanArray(cols) } //indicates if a solid boundary is present
 
     //for UI purposes; not used in calculations
-    private val density: Array<FloatArray> = Array(rows) { FloatArray(cols) }
+    private val density: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
     //for UI purposes; not used in calculations
-    private val vSquared: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val vX: Array<FloatArray> = Array(rows) { FloatArray(cols) }
-    private val vY: Array<FloatArray> = Array(rows) { FloatArray(cols) }
+    private val vSquared: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val vX: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
+    private val vY: Array<DoubleArray> = Array(rows) { DoubleArray(cols) }
 
     init {
         // initiate fluid with discretized velocity vectors
@@ -34,30 +34,30 @@ class BoltzmannFluid(
         // it does not matter if element is solid as the solver disregards velocity data
         // lots of values are precalculated here to make sure no unnecessary operations are performed
 
-        val here = 4.0f / 9.0f * (1 - 1.5f * overallVelocity * overallVelocity)
-        val up = 1.0f / 9.0f * (1 - 1.5f * overallVelocity * overallVelocity)
-        val down = 1.0f / 9.0f * (1 - 1.5f * overallVelocity * overallVelocity)
-        val left = 1.0f / 9.0f * (1 - 3.0f * overallVelocity + 3.0f * overallVelocity * overallVelocity)
-        val right = 1.0f / 9.0f * (1 + 3.0f * overallVelocity + 3.0f * overallVelocity * overallVelocity)
-        val northEast = 1.0f / 36.0f * (1 + 3.0f * overallVelocity + 3.0f * overallVelocity * overallVelocity)
-        val northWest = 1.0f / 36.0f * (1 - 3.0f * overallVelocity + 3.0f * overallVelocity * overallVelocity)
-        val southEast = 1.0f / 36.0f * (1 + 3.0f * overallVelocity + 3.0f * overallVelocity * overallVelocity)
-        val southWest = 1.0f / 36.0f * (1 - 3.0f * overallVelocity + 3.0f * overallVelocity * overallVelocity)
+        val here = 4.0 / 9.0 * (1 - 1.5 * overallVelocity * overallVelocity)
+        val up = 1.0 / 9.0 * (1 - 1.5 * overallVelocity * overallVelocity)
+        val down = 1.0 / 9.0 * (1 - 1.5 * overallVelocity * overallVelocity)
+        val left = 1.0 / 9.0 * (1 - 3.0 * overallVelocity + 3.0 * overallVelocity * overallVelocity)
+        val right = 1.0 / 9.0 * (1 + 3.0 * overallVelocity + 3.0 * overallVelocity * overallVelocity)
+        val northEast = 1.0 / 36.0 * (1 + 3.0 * overallVelocity + 3.0 * overallVelocity * overallVelocity)
+        val northWest = 1.0 / 36.0 * (1 - 3.0 * overallVelocity + 3.0 * overallVelocity * overallVelocity)
+        val southEast = 1.0 / 36.0 * (1 + 3.0 * overallVelocity + 3.0 * overallVelocity * overallVelocity)
+        val southWest = 1.0 / 36.0 * (1 - 3.0 * overallVelocity + 3.0 * overallVelocity * overallVelocity)
 
         for (r in 0..<rows) {
             for (c in 0..<cols) {
                 if (solid[r][c]) {
                     // clear the values
-                    velocityHere[r][c] = 0.0f
-                    velocityRight[r][c] = 0.0f
-                    velocityLeft[r][c] = 0.0f
-                    velocityUp[r][c] = 0.0f
-                    velocityDown[r][c] = 0.0f
-                    velocityNorthEast[r][c] = 0.0f
-                    velocityNorthWest[r][c] = 0.0f
-                    velocitySouthEast[r][c] = 0.0f
-                    velocitySouthWest[r][c] = 0.0f
-                    vSquared[r][c] = 0.0f
+                    velocityHere[r][c] = 0.0
+                    velocityRight[r][c] = 0.0
+                    velocityLeft[r][c] = 0.0
+                    velocityUp[r][c] = 0.0
+                    velocityDown[r][c] = 0.0
+                    velocityNorthEast[r][c] = 0.0
+                    velocityNorthWest[r][c] = 0.0
+                    velocitySouthEast[r][c] = 0.0
+                    velocitySouthWest[r][c] = 0.0
+                    vSquared[r][c] = 0.0
                 } else {
                     // init values for each element
                     velocityHere[r][c] = here
@@ -69,7 +69,7 @@ class BoltzmannFluid(
                     velocitySouthEast[r][c] = southEast
                     velocityNorthWest[r][c] = northWest
                     velocitySouthWest[r][c] = southWest
-                    density[r][c] = 1.0f
+                    density[r][c] = 1.0
                     vSquared[r][c] = overallVelocity * overallVelocity
                 }
             }
@@ -86,8 +86,8 @@ class BoltzmannFluid(
     }
 
     private fun collide() {
-        var sumVelocities: Float
-        val relaxationTime = 1 / (3f * viscosity + 0.5f) // omega in the equation; mostly an experimental value. Tweak for different results.
+        var sumVelocities: Double
+        val relaxationTime = 1 / (3.0 * viscosity + 0.5) // omega in the equation; mostly an experimental value. Tweak for different results.
 
         for (x in 0..<rows) {
             for (y in 0..<cols) {
@@ -97,8 +97,8 @@ class BoltzmannFluid(
                         + velocitySouthEast[x][y])
 
                     density[x][y] = sumVelocities //sets total density rho for UI
-                    val xVelocity: Float //flow velocity in x basis
-                    val yVelocity: Float //flow velocity in y basis
+                    val xVelocity: Double //flow velocity in x basis
+                    val yVelocity: Double //flow velocity in y basis
 
                     // calculate the flow velocities
                     if (sumVelocities > 0) {
@@ -109,14 +109,14 @@ class BoltzmannFluid(
                             + velocityNorthWest[x][y]) - velocityDown[x][y]
                             - velocitySouthEast[x][y] - velocitySouthWest[x][y])) / sumVelocities
                     } else {
-                        xVelocity = 0.0f
-                        yVelocity = 0.0f
+                        xVelocity = 0.0
+                        yVelocity = 0.0
                     }
 
 
                     // constants for the following calculations, different for every iteration of the loop
-                    val threeXVelocity = 3.0f * xVelocity
-                    val threeYVelocity = 3.0f * yVelocity
+                    val threeXVelocity = 3.0 * xVelocity
+                    val threeYVelocity = 3.0 * yVelocity
                     val xVelocitySquared = xVelocity * xVelocity
                     val yVelocitySquared = yVelocity * yVelocity
                     val twoXYVelocities = 2 * xVelocity * yVelocity
@@ -129,25 +129,25 @@ class BoltzmannFluid(
                     // Easy replacement for the magnitude of the curl, which is more difficult to calculate
 
                     // sets thermal velocities after collision
-                    velocityHere[x][y] += relaxationTime * (((4.0f / 9.0f) * sumVelocities
-                        * (1 - 1.5f * sumSquares)) - velocityHere[x][y])
-                    velocityRight[x][y] += relaxationTime * (((1.0f / 9.0f) * sumVelocities
-                        * (1 + threeXVelocity + 4.5f * xVelocitySquared - 1.5f * sumSquares)) - velocityRight[x][y])
-                    velocityLeft[x][y] += relaxationTime * (((1.0f / 9.0f) * sumVelocities
-                        * (1 - threeXVelocity + 4.5f * xVelocitySquared - 1.5f * sumSquares)) - velocityLeft[x][y])
-                    velocityUp[x][y] += relaxationTime * (((1.0f / 9.0f) * sumVelocities
-                        * (1 + threeYVelocity + 4.5f * yVelocitySquared - 1.5f * sumSquares)) - velocityUp[x][y])
-                    velocityDown[x][y] += relaxationTime * (((1.0f / 9.0f) * sumVelocities
-                        * (1 - threeYVelocity + 4.5f * yVelocitySquared - 1.5f * sumSquares)) - velocityDown[x][y])
+                    velocityHere[x][y] += relaxationTime * (((4.0 / 9.0) * sumVelocities
+                        * (1 - 1.5 * sumSquares)) - velocityHere[x][y])
+                    velocityRight[x][y] += relaxationTime * (((1.0 / 9.0) * sumVelocities
+                        * (1 + threeXVelocity + 4.5 * xVelocitySquared - 1.5 * sumSquares)) - velocityRight[x][y])
+                    velocityLeft[x][y] += relaxationTime * (((1.0 / 9.0) * sumVelocities
+                        * (1 - threeXVelocity + 4.5 * xVelocitySquared - 1.5 * sumSquares)) - velocityLeft[x][y])
+                    velocityUp[x][y] += relaxationTime * (((1.0 / 9.0) * sumVelocities
+                        * (1 + threeYVelocity + 4.5 * yVelocitySquared - 1.5 * sumSquares)) - velocityUp[x][y])
+                    velocityDown[x][y] += relaxationTime * (((1.0 / 9.0) * sumVelocities
+                        * (1 - threeYVelocity + 4.5 * yVelocitySquared - 1.5 * sumSquares)) - velocityDown[x][y])
 
-                    velocityNorthEast[x][y] += relaxationTime * (((1.0f / 36.0f) * sumVelocities
-                        * (1 + threeXVelocity + threeYVelocity + 4.5f * (sumSquares + twoXYVelocities) - 1.5f * sumSquares)) - velocityNorthEast[x][y])
-                    velocityNorthWest[x][y] += relaxationTime * (((1.0f / 36.0f) * sumVelocities
-                        * (1 - threeXVelocity + threeYVelocity + 4.5f * (sumSquares - twoXYVelocities) - 1.5f * sumSquares)) - velocityNorthWest[x][y])
-                    velocitySouthEast[x][y] += relaxationTime * (((1.0f / 36.0f) * sumVelocities
-                        * (1 + threeXVelocity - threeYVelocity + 4.5f * (sumSquares - twoXYVelocities) - 1.5f * sumSquares)) - velocitySouthEast[x][y])
-                    velocitySouthWest[x][y] += relaxationTime * (((1.0f / 36.0f) * sumVelocities
-                        * (1 - threeXVelocity - threeYVelocity + 4.5f * (sumSquares + twoXYVelocities) - 1.5f * sumSquares)) - velocitySouthWest[x][y])
+                    velocityNorthEast[x][y] += relaxationTime * (((1.0 / 36.0) * sumVelocities
+                        * (1 + threeXVelocity + threeYVelocity + 4.5 * (sumSquares + twoXYVelocities) - 1.5 * sumSquares)) - velocityNorthEast[x][y])
+                    velocityNorthWest[x][y] += relaxationTime * (((1.0 / 36.0) * sumVelocities
+                        * (1 - threeXVelocity + threeYVelocity + 4.5 * (sumSquares - twoXYVelocities) - 1.5 * sumSquares)) - velocityNorthWest[x][y])
+                    velocitySouthEast[x][y] += relaxationTime * (((1.0 / 36.0) * sumVelocities
+                        * (1 + threeXVelocity - threeYVelocity + 4.5 * (sumSquares - twoXYVelocities) - 1.5 * sumSquares)) - velocitySouthEast[x][y])
+                    velocitySouthWest[x][y] += relaxationTime * (((1.0 / 36.0) * sumVelocities
+                        * (1 - threeXVelocity - threeYVelocity + 4.5 * (sumSquares + twoXYVelocities) - 1.5 * sumSquares)) - velocitySouthWest[x][y])
                 }
             }
         }
@@ -162,36 +162,36 @@ class BoltzmannFluid(
                 if (solid[x][y]) {
                     if (velocityUp[x][y] > 0) {
                         velocityDown[x][y - 1] += velocityUp[x][y]
-                        velocityUp[x][y] = 0.0f
+                        velocityUp[x][y] = 0.0
                     }
                     if (velocityDown[x][y] > 0) {
                         velocityUp[x][y + 1] += velocityDown[x][y]
-                        velocityDown[x][y] = 0.0f
+                        velocityDown[x][y] = 0.0
                     }
                     if (velocityRight[x][y] > 0) {
                         velocityLeft[x - 1][y] += velocityRight[x][y]
-                        velocityRight[x][y] = 0.0f
+                        velocityRight[x][y] = 0.0
                     }
                     if (velocityLeft[x][y] > 0) {
                         velocityRight[x + 1][y] += velocityLeft[x][y]
-                        velocityLeft[x][y] = 0.0f
+                        velocityLeft[x][y] = 0.0
                     }
 
                     if (velocityNorthWest[x][y] > 0) {
                         velocitySouthEast[x + 1][y - 1] += velocityNorthWest[x][y]
-                        velocityNorthWest[x][y] = 0.0f
+                        velocityNorthWest[x][y] = 0.0
                     }
                     if (velocityNorthEast[x][y] > 0) {
                         velocitySouthWest[x - 1][y - 1] += velocityNorthEast[x][y]
-                        velocityNorthEast[x][y] = 0.0f
+                        velocityNorthEast[x][y] = 0.0
                     }
                     if (velocitySouthWest[x][y] > 0) {
                         velocityNorthEast[x + 1][y + 1] += velocitySouthWest[x][y]
-                        velocitySouthWest[x][y] = 0.0f
+                        velocitySouthWest[x][y] = 0.0
                     }
                     if (velocitySouthEast[x][y] > 0) {
                         velocityNorthWest[x - 1][y + 1] += velocitySouthEast[x][y]
-                        velocitySouthEast[x][y] = 0.0f
+                        velocitySouthEast[x][y] = 0.0
                     }
                 }
             }
@@ -206,17 +206,17 @@ class BoltzmannFluid(
         // constants for the following loops
 
         val v = overallVelocity
-        val threeTimesOverallVelocity = 3.0f * v
+        val threeTimesOverallVelocity = 3.0 * v
         val threeTimesOverallVelocitySquared = threeTimesOverallVelocity * v
-        val here = 4.0f / 9.0f * (1 - 1.5f * v * v)
-        val up = 1.0f / 9.0f * (1 - 1.5f * v * v)
-        val down = 1.0f / 9.0f * (1 - 1.5f * v * v)
-        val left = 1.0f / 9.0f * (1 - threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
-        val right = 1.0f / 9.0f * (1 + threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
-        val northEast = 1.0f / 36.0f * (1 + threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
-        val northWest = 1.0f / 36.0f * (1 - threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
-        val southEast = 1.0f / 36.0f * (1 + threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
-        val southWest = 1.0f / 36.0f * (1 - threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
+        val here = 4.0 / 9.0 * (1 - 1.5 * v * v)
+        val up = 1.0 / 9.0 * (1 - 1.5 * v * v)
+        val down = 1.0 / 9.0 * (1 - 1.5 * v * v)
+        val left = 1.0 / 9.0 * (1 - threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
+        val right = 1.0 / 9.0 * (1 + threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
+        val northEast = 1.0 / 36.0 * (1 + threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
+        val northWest = 1.0 / 36.0 * (1 - threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
+        val southEast = 1.0 / 36.0 * (1 + threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
+        val southWest = 1.0 / 36.0 * (1 - threeTimesOverallVelocity + threeTimesOverallVelocitySquared)
 
         // handle edges
         for (c in 0..<cols - 1) {
@@ -326,10 +326,10 @@ class BoltzmannFluid(
         consumer: (
             x: Int, y: Int,
             solid: Boolean,
-            density: Float,
-            vx: Float,
-            vy: Float,
-            velocity: Float,
+            density: Double,
+            vx: Double,
+            vy: Double,
+            velocity: Double,
         ) -> Unit
     ) {
         for (x in 0..<rows) {
